feat: fixup content watching and HMR
We now trigger (debounced) HMR when the .json files change, rather than
on mdx change. This is more correct (the mdx files only indirectly
affect the website content), and it also solves the issue of HMR not
working for gitignored mdx content (ie. the Mirascope v2 content that
gets pulled in externally).